Consolidate all components in src folder GH-386
