+++ .changeset/early-beans-impress.md
@@ -1,0 +1,5 @@
+---
+'@undataforum/components': patch
+---
+
+Consolidate all components in src folder
