+++ docs/src/components/EmailIcon.stories.js
@@ -1,21 +1,21 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { Box } from 'theme-ui';
-import { MenuIcon } from '@undataforum/components';
+import { EmailIcon } from '@undataforum/components';
 
-storiesOf('Icons/MenuIcon', module)
+storiesOf('Components/EmailIcon', module)
   .add('scale to height', () => (
     <Box sx={{ height: 64, m: 2 }}>
-      <MenuIcon scaleTo="height" />
+      <EmailIcon scaleTo="height" />
     </Box>
   ))
   .add('scale to width', () => (
     <Box sx={{ width: 64, m: 2 }}>
-      <MenuIcon scaleTo="width" />
+      <EmailIcon scaleTo="width" />
     </Box>
   ))
   .add('custom color', () => (
     <Box sx={{ height: 64, m: 2, color: 'primary' }}>
-      <MenuIcon scaleTo="height" />
+      <EmailIcon scaleTo="height" />
     </Box>
   ));
