+++ docs/src/components/GitHubIcon.stories.js
@@ -1,21 +1,21 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { Box } from 'theme-ui';
-import { TwitterIcon } from '@undataforum/components';
+import { GitHubIcon } from '@undataforum/components';
 
-storiesOf('Icons/TwitterIcon', module)
+storiesOf('Components/GitHubIcon', module)
   .add('scale to height', () => (
     <Box sx={{ height: 64, m: 2 }}>
-      <TwitterIcon scaleTo="height" />
+      <GitHubIcon scaleTo="height" />{' '}
     </Box>
   ))
   .add('scale to width', () => (
     <Box sx={{ width: 64, m: 2 }}>
-      <TwitterIcon scaleTo="width" />
+      <GitHubIcon scaleTo="width" />
     </Box>
   ))
   .add('custom color', () => (
     <Box sx={{ height: 64, m: 2, color: 'primary' }}>
-      <TwitterIcon scaleTo="height" />
+      <GitHubIcon scaleTo="height" />
     </Box>
   ));
