+++ docs/src/components/TwitterIcon.stories.js
@@ -1,21 +1,21 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { Box } from 'theme-ui';
-import { CloseIcon } from '@undataforum/components';
+import { TwitterIcon } from '@undataforum/components';
 
-storiesOf('Icons/CloseIcon', module)
+storiesOf('Components/TwitterIcon', module)
   .add('scale to height', () => (
     <Box sx={{ height: 64, m: 2 }}>
-      <CloseIcon scaleTo="height" />
+      <TwitterIcon scaleTo="height" />
     </Box>
   ))
   .add('scale to width', () => (
     <Box sx={{ width: 64, m: 2 }}>
-      <CloseIcon scaleTo="width" />
+      <TwitterIcon scaleTo="width" />
     </Box>
   ))
   .add('custom color', () => (
     <Box sx={{ height: 64, m: 2, color: 'primary' }}>
-      <CloseIcon scaleTo="height" />
+      <TwitterIcon scaleTo="height" />
     </Box>
   ));
