+++ packages/components/src/footer.js
