--- packages/components/src/icons/index.js
@@ -1,3 +1,0 @@
-export { default as EmailIcon } from './email-icon';
-export { default as GitHubIcon } from './github-icon';
-export { default as TwitterIcon } from './twitter-icon';
