+++ packages/components/src/post-preview.js
@@ -2,8 +2,7 @@ import React from 'react';
 import { func, shape, string } from 'prop-types';
 import { Button, Flex, Heading, Text } from 'theme-ui';
 
-import { responsiveNumberType } from '../types';
-
+import { responsiveNumberType } from './types';
 import SmartLink from './smart-link';
 
 const PostPreview = ({
