+++ packages/components/src/smart-link.js
